CA-46289: PIF_metrics vendor and devices names read incorrectly
PIF vendor and device names were listed as "unknown" in the PIF_metrics. This was due to a recent change in the Unixext.readfile_line function in combination with a slightly dubious way of jumping out of the read loop by raising End_of_file in pciutil.ml.

It is now possible to raise Unixext.Break in the function passed to Unixext.readfile_line, to break out of the read loop (if you really want to).

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>