CA-42914: Catch other exceptions when reading commands from client in block_device_io
Previously, we only caught End_of_file which Unixext.really_read throws when the client sends EOF. Other exceptions dribbled through to the deeper exception handler, which was supposed to be reserved exclusively for problems opening the block device.

Now, we also catch other exceptions in the same place as the End_of_file exception is handled.

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>