CA-42914, SCTX-434: Increase the timeout for writing the database to the redo-log
Previously, we had a flat timeout of 2 seconds for all redo-log operations. However, this has been shown to be too impatient for writing large databases over slow connections. (This resulted in timeouts firing prematurely, resulting in lots of METADATA_LUN_BROKEN alerts, and the redo-log being entirely useless for large pools as the database would never be successfully written!)

The timeout for each operation can now be specified independently. The new default for database writes is 30 seconds.

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>