CA-32077 Gracefully handle missing /etc/xensource-inventory file
xapi was failing nastily on start-up if the xensource-inventory file was
missing.  Now it generates a minimal one if none exists. This does not include
a build number, so version.ml now falls back to using a build number from the
Make environment if one is not available from the inventory, i.e. it falls
back to the behaviour from before Matthias's commit for CA-43574 (build number
from xensource-inventory).

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>