Call /opt/xensource/sm/mpathcount.py after xapi started if root disk is multipathed
mpathcount.py now updates the host object when the root disk is multipathed as well
as updating pbd objects.  (It writes values into other-config to show XenCenter
how many paths are active and how many are failed.)

Normally it is multipathd  that calls mpathcount.py, but in the case of the root
disk the /dev/mapper node is created by the initrd before multipathd is started.

Signed-off-by: Alex Zeffertt <alex.zeffertt@eu.citrix.com>