CA-47556: Revert to PV template for RHEL 6 64 bit for now
From ticket description:

See CP-1876 - RH have changed their kernel to require an option to use PV on HVM.
We currently have no way to set that option ourselves so would rely on the customer
doing it which they will be very unlikely to do in most cases. Also it's not easy
to switch from emulated drivers to PV therefore the chances are customers will end
up with a poorly performing fully HVM VM.

For RTM I suggest we revert to using the PV template for 64 bit (just like we do
for 32 bit). This is a simple revert of the template code.

Long term we want to move to HVM to avoid the performance penalty of 64 bit PV but
RH6 isn't quite ready yet. This could be a change we make when we turn RH6 support
into fully supported, rather than experimental, in a future LCM update

Signed-off-by: Alex Zeffertt <alex.zeffertt@eu.citrix.com>