Make xiu add a vnc-port entry in XenStore.
Starting simulated VMs when using xiu is completely broken at the moment.
xapi waits for a vnc-port to be written to XenStore (in one of two
places, depending upon whether the VM is HVM or PV).  xiu doesn't
currently write a value for this field, and so the VM start hangs for 20
minutes until the watch times out.

This patch writes a vnc-port entry in the two necessary places, for
every VM that is created.