CA-47144: PIFs were not recreated correctly after pool.eject
Problems arose when the slave had a plugged bond, especially if management was on the bond, but only when the bridging backend is in use.

The linux bonding driver sets the MAC addresses of the bond interface, as well as the slave interfaces, all to same MAC when the bond is up. The "real" MAC adresses of the slaves can be found in /proc/net/bonding/<bond if>. This patch fixes the Netdev.get_address function to take this into account.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>