CA-51022: Fix bug where after upgrade the old database is ignored and a fresh one created.
After upgrade the state.db will be present but the .generation file missing.
Consider a database to be present if only the state.db file is present
(defaulting the generation to 0) rather than requiring both files.

Note: this explains why an upgrade initially "worked": the fresh database
contains no records that need to be upgraded, masking the bug in the upgrade
code.

Note: we ought to continue the process of making the database self-contained
by placing the generation count into the .db file itself.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>