Check error code, instead of assuming failure.
parse_result_code of ocaml/xapi/workload_balancing.ml is called if
there is a result from a call to the WLB server.  In Windows, WLB's
Windows-Communication-Foundation web service returns no result when a
call is successful (obeying the WCF data contract).  When the same
code is run under Mono, the web service returns ResultCode = 0, which
XAPI is currently treating as an actual error.  The change is meant to
check the actual code rather than just assuming that since there is
one, that there was an error.

Brian Donegan sent in the original patch, that Matthias adapted.

Signed-off-by: Matthias Goergens <matthias.goergens@gmail.com>