CA-51925: Preserve the local-cache-sr field on pool-join
The local-cache-sr field was not being initialised on the new host object created on the master during pool-join.

Unfortunately, the only means of updating this field through the API is via Host.enable_local_storage_caching, which is an API call that is forwarded to the host in question. Unfortunately, during a pool-join, the host is offline so this mechanism cannot be used to set that field.

The solution is to make local-cache-sr a StaticRO field in the datamodel, so that it can be set in the constructor for the host object. This change to the API will not break backwards-compatibility since local-cache-sr is listed after all other StaticRO fields for the host object.

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>