CA-53000: Bug in xe-backup-metadata preventing scheduled metadata backups from working correctly
When the script is called with the -c option fs_uninitialised is set and it works the first time because there is no backup VDI. This variable is reset in the init_fs block which is executed only when a new backup VDI is created. Next time, a backup VDI exists and the init_fs block is not executed, fs_uninitialised remains set. During cleanup, if this variable is set, the VDI is destroyed.

The patched script resets this variable if create_vdi is set (-c option) and a VDI already exists.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>