CA-63739: Fix console handling after reboot/migrate/vncterm crash; add "xe console vm=..." for debugging
Previously if a client is watching the vncterm console of a VM while it reboots, then
a fresh console is created for the new domain but the old console lingers
while a connection remains open to it. This causes the user to see a "freeze"
and us to leak a vncterm and a proxy thread in xapi.

Now when the domain is destroyed xapi will kill the vncterm process group,
causing the client to disconnect and then reconnect to the real console
later, when it appears.

Previously When vncterm crashes, the local VNC port becomes free. A new vncterm or
qemu may then bind to the port. Two users may then accidentally get the same
console.

Now before opening a proxy to a particular local console port, we check that
the pid of vncterm or qemu is still valid, and reject the call with a 404
if not.

To help debugging it is nolonger necessary to port forward to vncterm at all.
Instead one may use 'xe' as follows:

$ xe console vm=<vm name or uuid>

The escape character is Control + ]

Note this only works with newer versions of the linux CLI and won't work
at all with the windows CLI. It remains a 'Hidden' command for now.

To make it easier to support CLI extensions in future the CLI protocol
version check has been loosened: now only the major versions need to match,
not the minor versions (oops).

Furthermore if one wants to experiment with PVFB, one only has to write:

$ xe vm-param-set uuid=... platform:pvfb=true

Signed-off-by: David Scott <dave.scott@eu.citrix.com>