Cache pool uuids contained in foreign databases.
When foreign database VDIs are added to the cache, keep a record of the
pool uuid of the foreign database, as well as the database generation
count. This allows VDI.metadata_of_pool to exposed to the CLI without
requiring the database to be opened every time the field is requested.

Signed-off-by: John Else <john.else@citrix.com>