Implement liveness monitor for metadata replication.
It would be good, in future, to unify this with the Redo_log_alert
functionality, i.e. enable multiple redo logs to be monitored and
generate system alerts. It is not possible to use the current HA
liveness monitor for this as it is designed to run indefinitely,
whereas we required monitoring to be turned on and off as redo_logs
are created and destroyed.

It should be possible, using the ocaml Event library, to listen for
accessibility state changes from all redo_logs at once using a single
long-running thread.

The monitor would then update the value of VDI.metadata_latest
accordingly via a callback - it is not possible to add this as a
callback to Redo_log.cannot_connect_fn as this would call the
redo_log database callback again, causing an infinite loop.

Signed-off-by: John Else <john.else@citrix.com>