CA-55979: Remove pre-unplugging of VIFs before bond.create
This led to some problems, so commenting out the code for now. We need to find a
better solution:

It would be great if we could simply move the interface from one bridge to
to the other, without involving the guest, so it would also work on guests that
do not support hot(un)plug of VIFs.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>