CA-48594: slightly improve the error handling of HTTP raw VDI upload
We delay sending back the HTTP headers until we've managed to vdi_attach
the disk, so that a storage error results in an HTTP error code as well as
a proper task error.

Note the most reliable way of monitoring error codes from HTTP calls is
to use a XenAPI task object.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>