Change VIF move function to move a dom0 vif device from one bridge to another
This function now no longer uses hot-unplug+plug, and should therefore
work for all guests. Some packet loss may occur during the move, but
for the current use cases (bond create/destroy) that is acceptable.

The patch adds a "move" option to the vif script, and adds to optional
parameters to the script for the domid and devid.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>