CA-56957: Make synchronising bonds on startup more synchronous
A Bond.create call that may be part of the "Synchronising bonds
on slave with master" startup task could continue even when the
task itself thought it stopped with an exception. We are now using
Xapi_bond.create directly rather than using an API call,
so that everything happens in the same thread.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>