CA-58406: Do not update MAC addresses of non-physical PIFs
MAC addresses of PIFs are synchronised with dom0 when xapi starts
to allow NICs to be replaced. Currently, xapi updates _all_ PIFs,
include bond and VLAN PIFs, which are not actual NICs. The MACs of
these non-physical NICs should not be updated. Especially in the
case of bond PIFs, as when the user creates a NIC bond, they are free
to choose any MAC they want, even if there is no NIC with a corresponding
MAC address in the system.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>