[CA-58880] Change VM_appliance.recover behaviour.
Previously when a VM appliance was recovered, if an appliance with the
same UUID already existed then a new appliance with a fresh UUID would
be created. In this case now, all VMs are removed from the existing
appliance, all recovered VMs are added and the appliance's name_label and
name_description are updated to match those of the recovered appliance.

Signed-off-by: John Else <john.else@citrix.com>