CA-59029: When calling sr_delete, the SM backends expect the SR to be attached *and* the PBD.currently_attached=true
In the storage interface, add SR.destroy and make it act the same way as SR.detach, except calling the underlying destroy function instead of the attach function. This means that it will also try to clean up any attached/activated VDIs before continuing.

It's not ideal that PBD.currently_attached=true because it's possible for other API calls to come in and touch the storage that we're about to delete.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>