To ensure backwards compatability, continue to use deprecated linux-specific encodings for the first 16 disks of HVM guests.
Ideally we would comply with the non-deprecated parts of the vbd-interface.txt in xen-4.1. However we have been using the linux-specific IDE encodings (hde..hdp) up until now; we will preserve this behaviour.

Note that the exact presentation to the guest depends on the behaviour of the PV drivers and the OS. In the case of Windows, the primary means of identifying disks is actually the volume label so even if devices got permuted on the IDE/xen bus it wouldn't matter.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>