CA-44742: Add a new API call 'event.from'
This is similar to the current 'event.next', but it doesn't require registration
before calling, and it doesn't lose create/modify events. Deletion events may get
lost. The return type is a structure containing the next token to pass in, the
list of events and the current valid reference counts. The latter is for the
verification of the client's cache - if the counts don't match after applying the
deltas received, the client should call 'get_all' on the relevant class in order
to remove any objects in the cache that have been deleted.

The first call to event.next can be used to populate the client's cache. Use the
empty string as a token, and all objects in the database will be returned.

Signed-off-by: Jon Ludlam <jonathan.ludlam@eu.citrix.com>