CA-44742: Add a couple of tests of the deleted-event code and change the semantics of fold_over_recent
fold_over_recent used to be inclusive - such that if you passed in a generation count of
n it would return events whose generation counts included n. Now it doesn't. This is due
to the use-case - the highest event number in the list returned is passed back in to the
fold_over_recent function, rather than the current generation count. The tests have been
changed to reflect the expected use of the API.

Signed-off-by: Jon Ludlam <jonathan.ludlam@eu.citrix.com>