CA-57168: Set limit of 1 vCPU for Ubuntu 10.04 in recommendations field
This is because of the
"kernel BUG at /build/buildd/linux-2.6.32/arch/x86/xen/spinlock.c:343!"
crash (which is most easily triggered by suspend/resume).

This bug can be fixed by updating the kernel (*), after which it will
still be possible to increase the number of vCPUs via the command line
since xe ignores the recommendations field.

(*) The Ubuntu 10.04 guest kernel can be updated to a natty backport
which does not have the issue like this:

  apt-get install python-software-properties # this installs add-apt-repository
  add-apt-repository ppa:kernel-ppa/ppa
  apt-get update
  apt-get install linux-headers-generic-pae-lts-backport-natty linux-image-generic-pae-lts-backport-natty # this installs 2.6.38.8.18
  echo xen_blkfront >> /etc/initramfs-tools/modules
  update-initramfs -c -k all # make sure xenblk module insmodded by initrd
  reboot

Signed-off-by: Alex Zeffertt <alex.zeffertt@eu.citrix.com>