CA-60198: Can't clear suspend-image-sr-uuid param on host
The generic "clear field" function tries to set the field to an empty string.
When clearing Host.suspend-image-sr-uuid, Host.clear-dump-sr-uuid, or
Pool.default-SR, the _set_ function tried to look-up an object with an empty
uuid, which caused the clear call to fail. This commit modifies these _set_
functions so that if an empty string is passed to it, the corresponding field
is set to Ref.null. This was already implemented for Pool.suspend-image-SR and
Pool.crash-dump-SR.

Signed-off-by: Rok Strnisa <rok.strnisa@citrix.com>