[CA-56359] Mark ignore_foreign_key fields as persistent.
When first populating its in-memory database, for each one-to-many
relationship xapi clears the field which is a set of references (i.e.
at the "one" end of the relationship) and regenerates these fields from
the "many" end of the relationship.

In the case of the role-subject relationship this fails, as there is no
Role.subject field with which to repopulate the Subject.roles field.

This change means that xapi no longer empties fields which are reference
sets if the field has ignore_foreign_key=true.

Signed-off-by: John Else <john.else@citrix.com>