[CA-49870] Allow xapi to start on slave with statefile access blocked.
This fixes the issue where if statefile access is blocked, all hosts in
a pool are shutdown, and a slave is restarted, the slave's xapi would
fail to start.

Sm.register makes a call to Sm_exec.make_call which attempts to look up
the local_cache_sr of the host, which fails as the master is not
contactable.

This change moves Sm.register later in the startup sequence, so that
xapi startup gets to the point where it can attempt to read the HA
statefile.

Signed-off-by: John Else <john.else@citrix.com>