[CA-49870] Enable slave to reboot under HA if statefile access blocked.
This change moves the slave's call to SM.register to a later position in
the startup sequence - in its previous position it would attempt to
make a database call to the master before the master connection had been
setup.

If HA was enabled, this database call would happen before the
slave had realised it needed to wait for statefile access, and xapi
would fail to start.

Signed-off-by: John Else <john.else@citrix.com>