CA-61774: Make Sync_networking more robust
Changing the {VLAN,Tunnel}.create calls in the slave resyncing
startup sequence to directly call the create functions (in Xapi_vlan etc.)
rather than using API calls (Client.VLAN.create etc). This was already the
case for the syncing of bonds.

The result is that the create calls are executed as part of the same thread
as the rest of the syncing code. Therefore, if an interface is changed inside
one of the create calls and the management connection is temporarily
interrupted, the rest of the sequence is not affected.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>