CA-61130: Remove unnecessary redo_log calls.
Redo_log.startup and Redo_log.flush_db_to_redo_log are unnecessary as
the new redo_log is already enabled so will be started up by the
database engine as soon as a database write occurs.

It is also dangerous to call them here, as the database engine can
request that all enabled redo_logs carry out a database flush at any
time - if this happens while a log is creating a block_device_io
process, we may end up with two processes, one of which the redo_log
will have lost track of.

This change means that the startup of the redo_log will be handled by
the database engine, so should not happen in two threads at once.

Signed-off-by: John Else <john.else@citrix.com>