CA-63225: Make qemu-dm-wrapper non-persistent
qemu-dm-wrapper changed in xen-api c/s 5d270b6 to fork-and-exec qemu-dm, so that the wrapper could forward stdout from qemu-dm to syslog. This entailed a larger dom0 memory overhead of roughly 1.5 MB per VM.

This patch reverts qemu-dm-wrapper to the old non-forking behaviour (whilst retaining the other changes in c/s 5d270b6). A subsequent patch will reinstate the forwarding of qemu-dm output to the syslog.

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>