CA-63225: Add fe option to output child process's stdout to syslog
When the syslog_stdout option is enabled, stdout from the child process is read line-by-line and output in real-time to syslog.

An optional argument (default: off) added to Forkhelpers.safe_close_and_exec and Forkhelpers.execute_command_get_output can be used to control this.

(Future work: This concept could be generalised to output any of the child process's open FDs to the syslog.)

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>