CA-61929: Fix metadata VDI handling callback
The metadata VDI handling callback which runs after an SR.scan called by
PBD.plug will now only run when the SR is shared and is being plugged to a
master.

If there is a scan of the SR in progress already, the callback will only
be called once the scan is complete.

Signed-off-by: John Else <john.else@citrix.com>