CA-64903: avoid "leaking" memory when a VM with no balloon driver is present on the system.
When a domain is starting, we allow is to use up to an "initial-reservation" (a soft limit managed by squeezed). Previously we considered the VM to have reached a stable memory state when the feature-balloon flag was written; this never happens if the guest has no balloon driver. We now consider the VM to have booted if either feature-balloon is written OR data/updated is written by the guest agent.

Once a domain has started, we set maxmem to the current total pages, releasing the unused parts of "initial-reservation".

Signed-off-by: David Scott <dave.scott@eu.citrix.com>