Add a version of init.d-xapissl for LSB platforms
This is a port of the original xapissl init.d script to work with Debian-based
systems. It should conform to the LSB standards for init scripts. I have
removed some weirdness from the original, such as its dependency on stunnelng,
hardcoding certain file paths, and storing conf files in odd places.

This commit also modifies the scripts Makefile to install this init script
alongside the original one, and tells the spec file to exclude this new one. In
the future, we should remove the old init script in favor of this one, but this
will serve as a holdover until then.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>