Optimise away unnecessary 'loopback' VBD.plugs: the storage layer vdi_attach will give us a block device already.
We use the convention that VBD.currently_attached = true iff the VDI is attached or activated by the storage layer.
Note that VBD.unplug will block (rather than fail with a transient) if a filehandle is still open; this is ok because dom0 VBDs are actively managed by xapi and unplugs are only attempted when the device is nolonger in use.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>