PR-1227: Do not create RRDs for bridges and VLAN PIFs
Currently, separate RRDs with read/write statistics are recorded with any network device,
which includes all PIFs including VLAN PIFs, all bridges, and all VIFs.

Each RRD takes up a certain amount of dom0 memory. With lots of VLANs configured in the system,
it is possible to run out of memory in dom0. Furthermore, a lot of dom0 CPU cycles are spent collecting
this data. ALso, clients that query the performance statistics (e.g. XenCenter when clicking on the
Performance tab), will face large downloads when lots of VLANs are present, even though VLAN stats
may not be displayed.

RRDs for VLANs in fact contain redundant information; the RRDs for the physical interfaces and VIFs
provide the same data. Likewise for the RRDs of bridges.

This patch removes VLAN and bridge RRDs in order to reduce the load on dom0's resources, especially
in the presence of lots of VLANs.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>