PR-1227: When creating a VLAN or Tunnel, reuse the PIF_metrics of the underlying PIF
Currently every VLAN PIF has its own PIF_metrics object. However, none of the metrics
actually make sense for individual VLAN PIFs, since most are properties of the physical PIF
(carrier, PCI information, speed, duplex). Therefore, for XS 6.0, we had made the carrier of a
VLAN PIF reflect the carrier of the underlying physical PIF. You could, however, consider doing
the same for all of the other metrics except the read/write stats.

The read/write stats are derived from the RRDs, which for VLANs have been removed in a previous patch,
and should therefore be removed as well.

A lot of dom0 resources are currently being wasted keeping all these metrics up to date. To improve
this situation, this patch replaces the individual metrics for VLAN PIFs by a single, shared PIF_metrics
object for a physical or bond PIF and all VLANs on top of it.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>