[PV PCI passthrough]: use Xenlight for PV PCI passthrough (but not HVM for now). This makes PV PCI passthrough work properly.
The xenops/ code for PV PCI passthrough is broken. At the very least it's
missing a call to xc_physdev_map_pirq, which itself is missing from the
Xenctrl library.

Since the xenlight PCI passthrough code is genuinely better than the xenops/
code, we should use it rather than fork it.

Note that PV PCI passthrough is still considered experimental, unlike the
HVM PCI passthrough codepath which is heavily used.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>