Tweak a few timeout values.
In principle, we'd like to set the timeouts to a little bit longer than the ordinary limit (just in case of stress),  but should definitely shorter than user's patience --- otherwise they would see the operations are hung and start to kill them anyway, or even worst, to kill or reboot the VMs or hosts out of panic. This is much worse than we setting and sticking to our own limit, failing quick and aborting the operation ourselves.

Signed-off-by: Zheng Li <zheng.li@eu.citrix.com>