[xenopsd]: use only 1 codepath for local VBD unplug
Previously there was one codepath for the 99% "normal" case where the unplug
is successful and a second 1% "transient failure" case where the unplug is
initially rejected, because of an unavoidable race in the kernel. The 1%
case was often broken and tricky to test. We resolve this by deliberately
triggering the "transient failure" case 100% of the time by opening a filehandle
ourselves, catching the initial failure and the closing the filehandle.