Xapi integration with xcp-networkd
All interaction with the networking daemon currently goes via the Nm
module, in the same way as it did for interface-reconfigure.
This is why the changes to xapi are concentrated in the Nm module.

It may be useful to review this in future, and have other modules
talk to the network daemon directly, for example to obtain information
about the network configuration.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>