CA-71731: when connection is reset, leave to the server side to figure out why
When xe transfer connection is reset, there are two possibilities: either a network issue, or some failure on server side. It's unfortunate that we can not distinguish the two cases with the current implementation, as shown in CA-71731. As a result, the client must always ask the server for a reason.

Signed-off-by: Zheng Li <zheng.li@eu.citrix.com>