Start xcp-networkd a little earlier in the boot sequence
It now starts at the same time as the old management-interface script.
This is to make sure the network is up for the scripts further down in the
startup sequence.

NOTE: the network daemon requires fe (xen-api-libs) to be running, so we need
to move fe up in the boot sequence before this patch can go in.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>