Improve ovs logging
OVS 1.2 added a log level called "off" that
is a better way to turn off logging, so we now use that in
interface-reconfigure.  It's a good idea to apply this, then, after
you update to OVS 1.2 or later.  There are no ill effects if you
don't, except that fatal errors (which wouldn't normally occur) will
be printed on the console as well as logged.

Signed-off-by: Ben Pfaff <blp@nicira.com>
Acked-by: Rob Hoes <rob.hoes@citrix.com>