networkd: set external-ids field on ovs bridges
Each ovs bridge is tagged with an "xs-network-uuids" key in its
external-ids map. The value of this key is set to the UUIDs of all
associated Network objects (separated by ";"), i.e. the network of
the physical or bond PIF plus the networks of all VLANs on top of it.

I don't actually know why this is needed, but I want to make sure that
networkd sets the same fields as interface-reconfigure to prevent anything
from breaking unexpectedly.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>