CA-63844: VBD.allowed_operations is inaccurate
VBD status report now shows that no operations are valid on a CD while its
VM is suspended.

When trying to insert/eject a CD while the VM is suspended, the error message
now shows that the VM should be in running _or shutdown_ state.

Currently, there are multiple places where checks are made about whether an
operation is permitted or not: at least where we are asking for allowed
operations, and where the operation is actually executed. All these places
should be combined in order to remove the possibility of de-synchronisation.

Signed-off-by: Rok Strnisa <rok.strnisa@citrix.com>