monitor: get stats/metrics for bond from underlying interface stats
...rather than from the linux bond interface. This is needed, because,
for the ovs, we want to get rid of the "fake" bond interface, which
was introduced only for compatility reasons (the interface is used only
for stats collection, and is no good for anything else).

This also gives us an opportunity to redefine the PIF_metrics for bonds
in xapi itself. The fields are now defined as follows:

* vendor_id, device_id: empty string
* speed: the sum of the speeds of the bonded interfaces
* duplex: the "minimum" duplex of the bonded interfaces
* carrier: logical OR of the carriers of the bonded interfaces
* pci_bus_path: empty string

The tx/rx metrics for the RRDs are the sums of these metrics of the
bonded interfaces.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>