CA-75873: Make sure HVM guests use the Windows check for PV driver up-to-date-ness
If a Windows guest is migrated or suspend/resumed then the key which
we use to check that it is Windows gets removed. This patch makes sure
that all HVM guests are treated as Windows, and so will use the build
number to determine whether drivers are out of date.

Signed-off-by: John Else <john.else@citrix.com>